created template for self moving arm
